Isolate vpnkit http server in a separate package
It’s required in order to avoid to vendor labstack/echo when using the vpnkit client.
